backend/requirements.txt
aerich==0.7.2
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
atlastk==0.13.3
click==8.1.7
ddlparse==1.10.0
dictdiffer==0.9.0
exceptiongroup==1.2.0
fastapi==0.109.0
h11==0.14.0
idna==3.6
iso8601==1.1.0
pydantic==2.5.3
pydantic-core==2.14.6
pyparsing==3.1.1
pypika-tortoise==0.1.6
python-dotenv==1.0.1
pytz==2023.3.post1
sniffio==1.3.0
starlette==0.35.1
tomlkit==0.12.3
tortoise==0.1.1
tortoise-orm==0.20.0
typing-extensions==4.9.0
uvicorn==0.27.0
